requirements.txt
absl-py==2.1.0
addict==2.4.0
albucore==0.0.13
albumentations==1.4.14
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
Django==5.1
easydict==1.13
eval_type_backport==0.2.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
grpcio==1.65.4
humanfriendly==10.0
idna==3.7
imageio==2.35.0
importlib_metadata==8.2.0
insightface==0.7.3
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.5.1
lpips==0.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
onnx==1.16.2
onnxruntime==1.18.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
prettytable==3.11.0
protobuf==5.27.3
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
tb-nightly==2.18.0a20240814
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.10
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
yapf==0.40.2
zipp==3.20.0
transformers
diffusers
